FE added code improvements
