+++ src/assets/main.scss
@@ -15,7 +15,7 @@ $color-black: #212121;
   --color-black: #{$color-black};
 }
 
-.toggle-knob-shadow {
+.ripple-shadow {
   opacity: 0;
   width: 30px;
   height: 30px;
@@ -29,4 +29,17 @@ $color-black: #212121;
     transform 0.2s ease-in-out;
   pointer-events: none;
   z-index: 2;
+}
+
+.body-option {
+  padding: 5px 0 5px 0;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+.widget-row {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0;
 }
