+++ src/types/widget.ts
@@ -1,9 +1,9 @@
 export interface Widget {
   id: number
-  type: string
+  type: 'carbon' | 'plastic bottles' | 'trees'
   amount: number
-  action: string
+  action: 'collects' | 'plants' | 'offsets'
   active: boolean
   linked: boolean
-  selectedColor: string
+  selectedColor: 'white' | 'black' | 'blue' | 'green' | 'beige'
 }
